:recycle: Refactoring to add a tpf module
I'm refactoring so that we have a subclass TPFMachine, to keep each module a bit more light weight. I suggest we in future make a subclass FFIMachine. It may be that we actually end up needing KeplerTPFMachine, K2TPFMachine, TESSTPFMachine...we'll see.
