+++ psfmachine/__init__.py
@@ -7,5 +7,5 @@ PACKAGEDIR = os.path.abspath(os.path.dirname(__file__))
 from .version import __version__
 
 # from .data import Cube
-from .collection import Collection
 from .machine import Machine
+from .tpf import TPFMachine
